Cambio de tipos de secretariado a las nuevas vista

+++ resources/views/tiposSecretariados/index.blade.php
@@ -4,11 +4,10 @@
 @endsection
 @section('contenido')
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
-            <div class="row">
-                @include('tiposSecretariados.parciales.buscar')
-            </div>
+            @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'tiposSecretariados.parciales.buscar'])
+            @include('comun.plantillaOperacionesIndex',['tabla'=>'tiposSecretariados','accion'=>'Nuevo'])
             @if(!$tipoSecretariados->isEmpty())
                 <div class="full-Width">
                     <table class="table-viaoptima table-striped">
@@ -33,24 +32,36 @@
                                             </i>
                                         </a>
                                         @if (Auth::user()->roles->peso>=config('opciones.roles.administrador'))
-                                            {!! FORM::open(array('route' => array('tiposSecretariados.destroy', $tipo_secretariado->id),
-                                            'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
-                                            <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
-                                                    @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                                    class="pull-right lanzarModal"
-                                                    data-title="BORRADO"
-                                                    data-descripcion="¿Seguro que deseas eliminar este tipo de secretariado?
+                                            @if ($tipo_secretariado->activo)
+                                                {!! FORM::open(array('route' => array('tiposSecretariados.destroy', $tipo_secretariado->id),
+                                                'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
+                                                <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
+                                                        @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                                        class="pull-right lanzarModal simpleModal"
+                                                        data-modal_sin_etiqueta="true"
+                                                        data-modal_ancho="330"
+                                                        data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
+                                                        data-modal_cabecera_color_texto='#ffffff'
+                                                        data-modal_cuerpo_color_fondo='rgba(255,255,255,.9)'
+                                                        data-modal_cuerpo_color_texto='"#ffffff'
+                                                        data-modal_pie_color_fondo='#400090'
+                                                        data-modal_pie_color_texto='"#ffffff'
+                                                        data-modal_posicion_vertical="220"
+                                                        data-titulo="BORRAR"
+                                                        data-pie="true"
+                                                        data-descripcion="¿Seguro que deseas eliminar este tipo de secretariado?
                                                     <h3><strong class='green'>{{ $tipo_secretariado->tipo_secretariado}}</strong></h3>"
-                                                    data-footer="true"
-                                                    @endif >
-                                                <i class='glyphicon glyphicon-trash full-Width'>
-                                                    <div>Borrar</div>
-                                                </i>
-                                            </button>
-                                            @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                                @include ("comun.plantillaBorrado")
+                                                        data-pie="true"
+                                                        @endif >
+                                                    <i class='glyphicon glyphicon-trash full-Width'>
+                                                        <div>Borrar</div>
+                                                    </i>
+                                                </button>
+                                                @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                                    @include ("comun.plantillaBorrado")
+                                                @endif
+                                                {!! FORM::close() !!}
                                             @endif
-                                            {!! FORM::close() !!}
                                         @endif
                                     </div>
                                 </td>
@@ -62,15 +73,12 @@
             @else
                 <div class="clearfix">
                     <div class="alert alert-info" role="alert">
-                        <p><strong>¡Aviso!</strong> No se ha encontrado ningun tipo de secretariado que listar.
+                        <p><strong>¡Aviso!</strong> No se ha encontrado ning&uacute;n tipo de secretariado que listar.
                         </p>
                     </div>
                 </div>
             @endif
-            <div class="row paginationBlock">
-                {!! $tipoSecretariados->appends(Request::only(['tipo_secretariado']))->render()
-                !!}{{-- Poner el paginador --}}
-            </div>
+            {!! $tipoSecretariados->appends(Request::only(['tipo_secretariado']))->render()!!}
         @else
             @include('comun.guestGoHome')
         @endif
